+++ auto_round/algorithms/quantization/awq/quantizer.py
@@ -60,7 +60,7 @@ from auto_round.utils import (
     set_amax_for_all_moe_layers,
     set_module,
 )
-from auto_round.utils.device_manager import device_manager
+from auto_round.devices.device_manager_haha import device_manager
 from auto_round.wrapper import WrapperLinear
 from auto_round.wrapper import WrapperMultiblock as _WrapperMultiblock
