+++ auto_round/calibration/inputs.py
@@ -18,7 +18,7 @@ from typing import Tuple
 import torch
 
 from auto_round.utils import clear_memory, to_device, to_dtype
-from auto_round.utils.device_manager import device_manager
+from auto_round.devices.device_manager_haha import device_manager
 
 __all__ = ["split_inputs", "preprocess_block_inputs"]
