+++ auto_round/devices/utils.py
@@ -13,17 +13,14 @@
 # limitations under the License.
 import ctypes
 import functools
-import gc
 import os
 import re
-import shutil
 import sys
-import tempfile
 from contextlib import ContextDecorator, contextmanager
 from functools import lru_cache
 from itertools import combinations
 from threading import Lock
-from typing import Any, Callable, Optional, Union
+from typing import Callable, Optional, Union
 
 import cpuinfo
 import psutil
@@ -32,12 +29,9 @@ from accelerate import dispatch_model, infer_auto_device_map
 from accelerate.utils import get_balanced_memory, get_max_memory
 
 from auto_round.logger import logger
-from auto_round.utils.device_manager import (
+from auto_round.devices.device_manager_haha import (
     clear_memory,
-    detect_device_count,
-    get_ar_device,
-    get_available_device_types,
-    get_current_device_manager,
+    get_current_ar_device,
     get_device_memory,
     get_major_device,
 )
@@ -58,234 +52,8 @@ DEVICE_ENVIRON_VARIABLE_MAPPING = {
 # 2. Lazy Mode (By default)
 
 
-################ Check available sys.module to decide behavior #################
-def is_package_available(package_name: str) -> bool:
-    """Check if the package exists in the environment without importing.
-
-    Args:
-        package_name (str): package name
-    """
-    from importlib.util import find_spec
-
-    package_spec = find_spec(package_name)
-    return package_spec is not None
 
 
-def is_hpu_lazy_mode():
-    return os.getenv("PT_HPU_LAZY_MODE") != "0"
-
-
-def _use_hpu_compile_mode():
-    from auto_round.utils.common import TORCH_VERSION_AT_LEAST_2_4
-
-    return TORCH_VERSION_AT_LEAST_2_4 and not is_hpu_lazy_mode()
-
-
-def _bump_dynamo_cache_limit(min_size: Optional[int] = None):
-    """Raise torch._dynamo cache/recompile limits.
-
-    The same quant function (e.g. ``quant_tensor_sym``) is reused across
-    every linear layer in a transformer block (q/k/v/o_proj, gate/up/
-    down_proj, ...), each with a different weight shape. Because dynamo's
-    compile cache is keyed by the function's code object (shared across
-    all WrapperLinear instances), per-layer static recompiles quickly
-    exceed the default ``recompile_limit`` (8) and trigger a fallback to
-    eager with a noisy warning. We keep static-shape compilation (best
-    perf) and just allow more cache entries.
-
-    The threshold can be overridden via the ``AR_DYNAMO_CACHE_SIZE_LIMIT``
-    environment variable (default: 16).
-    """
-    try:
-        if min_size is None:
-            from auto_round import envs
-
-            min_size = envs.AR_DYNAMO_CACHE_SIZE_LIMIT
-        from torch._dynamo import config as _dynamo_config
-
-        for attr in ("cache_size_limit", "accumulated_cache_size_limit", "recompile_limit"):
-            if hasattr(_dynamo_config, attr) and getattr(_dynamo_config, attr) < min_size:
-                setattr(_dynamo_config, attr, min_size)
-    except Exception:  # pragma: no cover - best effort
-        pass
-
-
-def compile_func(
-    fun: Union[torch.nn.Module, Callable], device: Union[str, torch.device, int]
-) -> Union[torch.nn.Module, Callable]:
-    """Compile a function on the specified device.
-
-    The shared dynamo cache-limit bump lives in :func:`_bump_dynamo_cache_limit`;
-    the per-device ``torch.compile`` customization (whether to compile and which
-    backend to use) is delegated to the corresponding :class:`Device`, keeping
-    this entry point device-agnostic.
-    """
-    return get_ar_device(device).compile_func(fun)
-
-
-def is_numba_available():  # pragma: no cover
-    """Check if Numba is available."""
-    try:
-        import numba
-
-        return True
-    except ImportError:
-        return False
-
-
-def _is_tbb_installed():  # pragma: no cover
-    import importlib.metadata
-
-    try:
-        importlib.metadata.version("tbb")
-        return True
-    except importlib.metadata.PackageNotFoundError:
-        return False
-
-
-def _is_tbb_configured():  # pragma: no cover
-    try:
-        from numba.np.ufunc.parallel import _check_tbb_version_compatible
-
-        # check if TBB is present and compatible
-        _check_tbb_version_compatible()
-
-        return True
-    except ImportError as e:
-        logger.warning_once(f"TBB not available: {e}")
-        return False
-
-
-def is_tbb_available():  # pragma: no cover
-    """Check if TBB is available."""
-    if not _is_tbb_installed():
-        logger.warning_once("TBB is not installed, please install it with `pip install tbb`.")
-        return False
-    if not _is_tbb_configured():
-        logger.warning_once(
-            (
-                "TBB is installed but not configured correctly. \n"
-                "Please add the TBB library path to `LD_LIBRARY_PATH`, "
-                "for example: `export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib/`."
-            )
-        )
-        return False
-    return True
-
-
-def can_pack_with_numba():  # pragma: no cover
-    """Check if Numba and TBB are available for packing.
-
-    To pack tensor with Numba, both Numba and TBB are required, and TBB should be configured correctly.
-    """
-    if not is_numba_available():
-        logger.warning_once("Numba is not installed, please install it with `pip install numba`.")
-        return False
-    if not is_tbb_available():
-        return False
-    return True
-
-
-## check hpex
-if is_package_available("habana_frameworks"):
-    _hpex_available = True
-    import habana_frameworks.torch.hpex  # pylint: disable=E0401
-else:
-    _hpex_available = False
-
-
-@torch._dynamo.disable()
-@lru_cache(None)
-def is_hpex_available():
-    return _hpex_available
-
-
-_xpu_sdpa_patched = False
-
-
-# TODO: This is a workaround for the XPU SDPA memory blow-up issue.
-# We should remove this patch after the issue is fixed in XPU side.
-# https://github.com/intel/auto-round/issues/990
-def patch_xpu_sdpa_drop_causal_mask():
-    """Workaround for XPU peak-VRAM blow-up in SDPA.
-
-    On Intel XPU, ``torch.nn.functional.scaled_dot_product_attention`` falls back
-    to the MATH backend whenever ``attn_mask`` is non-None (FLASH on XPU does
-    not support explicit attn_mask, EFFICIENT/CUDNN are not implemented).
-    The MATH backend materializes the full ``(B, H, S, S)`` probability matrix
-    in both forward and backward, costing several GB at typical
-    ``batch_size=8, seqlen=2048``.
-
-    HuggingFace transformers happily passes a *pure causal* 4D mask, even
-    though the same effect is achievable via ``is_causal=True`` (which the
-    FLASH backend supports and which uses ~12x less memory).
-
-    This monkey-patch detects a pure causal mask, drops it, and forwards
-    ``is_causal=True`` instead -- only on XPU and only when no real mask info
-    would be lost. CPU/CUDA/HPU paths are left untouched.
-
-    Idempotent. Safe to call multiple times.
-    """
-    global _xpu_sdpa_patched
-    if _xpu_sdpa_patched:
-        return
-    if not (hasattr(torch, "xpu") and torch.xpu.is_available()):
-        return
-    torch.use_deterministic_algorithms(False)
-
-    _orig_sdpa = torch.nn.functional.scaled_dot_product_attention
-
-    def _is_pure_causal_mask(mask: torch.Tensor, s: int) -> bool:
-        # Cheap shape check first
-        if mask.shape[-1] != s or mask.shape[-2] != s:
-            return False
-        if mask.dtype not in (torch.float16, torch.bfloat16, torch.float32):
-            return False
-        # Pick the first (B,H) slice; HF mask is broadcast across batch/heads.
-        m2d = mask.reshape(-1, s, s)[0]
-        tri_up = torch.triu(torch.ones(s, s, dtype=torch.bool, device=mask.device), 1)
-        return bool(torch.isinf(m2d[tri_up]).all().item()) and bool((m2d[~tri_up] == 0).all().item())
-
-    def _patched_sdpa(query, key, value, attn_mask=None, dropout_p=0.0, is_causal=False, scale=None, **kwargs):
-        if (
-            query.device.type == "xpu"
-            and attn_mask is not None
-            and not is_causal
-            and query.shape[-2] == key.shape[-2]  # square QK length (no kv-cache)
-            and _is_pure_causal_mask(attn_mask, query.shape[-2])
-        ):
-            attn_mask = None
-            is_causal = True
-        return _orig_sdpa(
-            query,
-            key,
-            value,
-            attn_mask=attn_mask,
-            dropout_p=dropout_p,
-            is_causal=is_causal,
-            scale=scale,
-            **kwargs,
-        )
-
-    torch.nn.functional.scaled_dot_product_attention = _patched_sdpa
-    _xpu_sdpa_patched = True
-    logger.info("torch.use_deterministic_algorithms(False) is set for XPU.")
-    logger.info(
-        "Patched torch SDPA on XPU to use is_causal=True for pure causal masks "
-        "(avoids ~10x peak-VRAM blow-up from MATH backend)."
-    )
-
-
-def check_is_cpu(device):
-    """Check if the device is a CPU.
-
-    Args:
-        device: The device to be checked.
-
-    Returns:
-        bool: True if the device is a CPU, False otherwise.
-    """
-    return device == torch.device("cpu") or device == "cpu"
 
 
 def is_pipeline_parallel_supported(device_type: str) -> bool:
@@ -515,7 +283,7 @@ def clear_memory_if_reached_threshold(threshold=0.85, device_list=None):
         bool: True if memory was cleared, False otherwise.
     """
     # Detect the active device (CUDA/XPU/HPU/...)
-    dev_mgr = get_current_device_manager()
+    dev_mgr = get_current_ar_device()
     if not dev_mgr.is_available() or dev_mgr.type == "cpu":
         return False
     name = dev_mgr.type
@@ -543,44 +311,6 @@ def clear_memory_if_reached_threshold(threshold=0.85, device_list=None):
             logger.warning_once(f"Failed to check memory for {name}:{i}: {e}")
     return False
 
-
-def check_memory_availability(device, inputs, weight, org_seqlen, org_bs):
-    """Checks the availability of memory on the specified device for processing inputs using a given weight tensor.
-
-    Args:
-        device (str): The device type ('cuda' for GPU or 'hpu' for HPU).
-        inputs (torch.Tensor): Input tensor.
-        weight (torch.Tensor): Weight tensor.
-        org_seqlen (int): Original sequence length.
-        org_bs (int): Original batch size.
-
-    Returns:
-        tuple: A tuple containing availability status (bool), modified sequence length (int),
-                and modified batch size (int).
-    """
-    weight_memory = weight.numel() * weight.element_size()
-    device_type = str(device).split(":")[0]
-    if device_type in ("cpu", "") or not get_ar_device(device_type).is_available():
-        return True, org_seqlen, org_bs
-    dev_mgr = get_ar_device(device_type)
-    current_index = dev_mgr.current_device()
-    free_space, _ = dev_mgr.mem_get_info(current_index)
-
-    free_space = free_space - weight_memory * 10  # for min_max_scale & grad usage
-    seqlen = org_seqlen
-    bs = org_bs
-    in_feature = weight.shape[1]
-    out_feature = weight.shape[0]
-    while seqlen >= 128:
-        input_size = bs * seqlen * in_feature
-        output_size = bs * seqlen * out_feature
-        input_output_memory = 2 * (input_size * inputs.element_size() + output_size * inputs.element_size())
-        if input_output_memory < free_space:
-            return True, seqlen, bs
-        seqlen = seqlen // 2
-        bs = 1
-
-    return False, seqlen, bs
 
 
 def set_tuning_device_for_layer(model, name: str, device: str) -> None:
@@ -834,7 +564,7 @@ def estimate_tuning_block_mem(
             - additional_memory (float): Additional memory overhead (in GB) for operations like attention.
     """
     # Calculate all block parameters memory and build layer-wise memory dict
-    from auto_round.utils.model import get_layer_features, is_moe_layer
+    from auto_round.utils.model import get_layer_features
 
     layer_memory_dict = {}
 
@@ -943,7 +673,7 @@ def set_auto_device_map_for_block_with_tuning(
         This function is intended for internal use in device memory management and tuning.
     """
     card_0_in_high_risk, loss_device = False, output_device
-    dev_mgr = get_current_device_manager()
+    dev_mgr = get_current_ar_device()
     if dev_mgr.is_available() and dev_mgr.type != "cpu":
         num_devices = dev_mgr.device_count()
         device_name = dev_mgr.type
@@ -1167,30 +897,6 @@ def set_avg_auto_device_map(model: torch.nn.Module, device_map):
                 device_index += 1
 
 
-if __name__ == "__main__":
-    # Example usage
-    number_dict = {"item1": 90, "item2": 20, "item3": 30, "item4": 40, "item5": 50, "item6": 60}
-
-    groups = partition_dict_numbers(number_dict, 10)
-    for i, group in enumerate(groups):
-        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
-
-    groups = partition_dict_numbers(number_dict, 6)
-    for i, group in enumerate(groups):
-        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
-
-    groups = partition_dict_numbers(number_dict, 4)
-    for i, group in enumerate(groups):
-        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
-
-    groups = partition_dict_numbers(number_dict, 3)
-    for i, group in enumerate(groups):
-        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
-
-    groups = partition_dict_numbers(number_dict, 2)
-    for i, group in enumerate(groups):
-        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
-
 
 def parse_available_devices(device_map: Union[str, torch.device, int, dict, None] = None) -> list:
     """
@@ -1209,25 +915,12 @@ def parse_available_devices(device_map: Union[str, torch.device, int, dict, None
     Returns:
         list[str]: Normalized device names, e.g., ["cuda:0", "cuda:1"] or ["cpu"]
     """
-
-    # === Step 1. Detect available device types ===
-    device_types = get_available_device_types()
-
-    # Always include CPU as a fallback
-    if not device_types:
-        device_types = ["cpu"]
-
+    device_types = get_major_device()
     # === Step 2. Parse different input formats ===
     if device_map is None:
         # Automatically detect one available device
-        if "cuda" in device_types:
-            return ["cuda:0"]
-        elif "xpu" in device_types:
-            return ["xpu:0"]
-        elif "hpu" in device_types:
-            return ["hpu:0"]
-        else:
-            return ["cpu"]
+        return [get_major_device()]
+
 
     if isinstance(device_map, torch.device):
         # Handle torch.device objects
@@ -1266,14 +959,11 @@ def parse_available_devices(device_map: Union[str, torch.device, int, dict, None
             return ["cpu"]
         if device_map.lower() == "auto":
             device_count = detect_device_count()
-            if "cuda" in device_types:
-                return [f"cuda:{i}" for i in range(device_count)]
-            elif "xpu" in device_types:
-                return [f"xpu:{i}" for i in range(device_count)]
-            elif "hpu" in device_types:
-                return [f"hpu:{i}" for i in range(device_count)]
+            if device_count == 0:
+                return device_types
             else:
-                return ["cpu"]
+                return [f"{device_types}:{i}" for i in range(device_count)]
+
         # Split by commas
         parts = [x.strip() for x in device_map.split(",") if x.strip()]
         parsed = []
@@ -1351,13 +1041,13 @@ class MemoryMonitor:
             if not isinstance(device_list, (list, tuple)):
                 device_list = [device_list]
         else:
-            dev_mgr = get_current_device_manager()
+            dev_mgr = get_current_ar_device()
             if dev_mgr.is_available() and dev_mgr.type != "cpu":
                 device_list = list(range(dev_mgr.device_count()))
             else:
                 device_list = [0]
 
-        dev_mgr = get_current_device_manager()
+        dev_mgr = get_current_ar_device()
         for device in device_list:
             if str(device) == "cpu":
                 continue
@@ -1653,3 +1343,28 @@ def dispatch_model_by_all_available_devices(
     device_map = infer_auto_device_map(model, max_memory=new_max_memory, no_split_module_classes=no_split_modules)
     model = dispatch_model(model, device_map=device_map)
     return model
+
+
+if __name__ == "__main__":
+    # Example usage
+    number_dict = {"item1": 90, "item2": 20, "item3": 30, "item4": 40, "item5": 50, "item6": 60}
+
+    groups = partition_dict_numbers(number_dict, 10)
+    for i, group in enumerate(groups):
+        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
+
+    groups = partition_dict_numbers(number_dict, 6)
+    for i, group in enumerate(groups):
+        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
+
+    groups = partition_dict_numbers(number_dict, 4)
+    for i, group in enumerate(groups):
+        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
+
+    groups = partition_dict_numbers(number_dict, 3)
+    for i, group in enumerate(groups):
+        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
+
+    groups = partition_dict_numbers(number_dict, 2)
+    for i, group in enumerate(groups):
+        print(f"Group {i + 1}: {group}, Sum: {sum(group.values())}")
