+++ auto_round/eval/evaluation.py
@@ -438,7 +438,7 @@ def run_model_evaluation(model, tokenizer, autoround, folders, formats, args):
 
     if args.tasks is None or args.tasks == "" or eval_folder is None:
         return
-    from auto_round.utils.device_manager import device_manager, get_device_and_parallelism
+    from auto_round.devices.device_manager_haha import device_manager, get_device_and_parallelism
 
     device_str, _ = get_device_and_parallelism(device_manager.device_map)
     # Handle vllm backend evaluation
